Added random generation project
This is a Intellij java project for developing the random generation. This is to allow us to work on developing random map/house/dialogue generation without integrating it into the Godot project and without waiting for the Godot project to be developed enough to add those features. It uses Java so that we can easily copy and paste it into Godot.